refactor: moved functions versions to core
