+++ src/cli/commands/init.ts
@@ -2,7 +2,6 @@ import chalk from "chalk";
 import { Command } from "commander";
 import path from "path";
 import process from "process";
-import prompts from "prompts";
 import { promptOrUseDefault } from "../../core/prompts";
 import {
   configureOptions,
@@ -15,6 +14,7 @@ import {
   writeConfigFile,
 } from "../../core/utils";
 import { detectProjectFolders, generateConfiguration, isDescendantPath } from "../../core/frameworks";
+import { getChoicesForApiLanguage } from "../../core/functions-versions";
 
 export default function registerCommand(program: Command) {
   program
@@ -203,6 +203,23 @@ async function promptConfigSettings(disablePrompts: boolean, detectedConfig: Fra
       format: trimValue,
     },
     {
+      type: (prev) => (prev != null ? "select" : null),
+      name: "apiLanguage",
+      message: "What's your API language? (optional)",
+      choices: [
+        { title: "Node.js", value: "node" },
+        { title: "Python", value: "python" },
+        { title: "Dotnet", value: "dotnet" },
+        { title: "Dotnet isolated", value: "dotnet-isolated" },
+      ],
+    },
+    {
+      type: (prev) => (prev != null ? "select" : null),
+      name: "apiVersion",
+      message: "What's your API version? (optional)",
+      choices: (prev) => getChoicesForApiLanguage(prev),
+    },
+    {
       type: "text",
       name: "appBuildCommand",
       message: "What command do you use to build your app? (optional)",
@@ -240,73 +257,7 @@ async function promptConfigSettings(disablePrompts: boolean, detectedConfig: Fra
     },
   ]);
 
-  if (response.apiLocation) {
-    const apiLanguagePrompt = await prompts([
-      {
-        type: "select",
-        name: "apiLanguage",
-        message: "What's your API language? (optional)",
-        choices: [
-          { title: "Node.js", value: "node" },
-          { title: "Python", value: "python" },
-          { title: "Dotnet", value: "dotnet" },
-          { title: "Dotnet isolated", value: "dotnet-isolated" },
-        ],
-      },
-    ]);
-
-    response.apiLanguage = apiLanguagePrompt.apiLanguage;
-
-    const apiVersionPrompt = await prompts([
-      {
-        type: "select",
-        name: "apiVersion",
-        message: "What's your API version? (optional)",
-        choices: getChoicesForApiLanguage(response.apiLanguage),
-      },
-    ]);
-
-    response.apiVersion = apiVersionPrompt.apiVersion;
-  }
-
   return response;
-}
-
-function getChoicesForApiLanguage(apiLanguage: string) {
-  // Refer to this for functions and versions - https://learn.microsoft.com/en-us/azure/static-web-apps/configuration#selecting-the-api-language-runtime-version
-  let choices = [];
-  switch (apiLanguage) {
-    case "node":
-      choices = [
-        { title: "16", value: "16" },
-        { title: "14", value: "14" },
-        { title: "12", value: "12" },
-      ];
-      break;
-    case "python":
-      choices = [
-        { title: "3.8", value: "3.8" },
-        { title: "3.9", value: "3.9" },
-      ];
-      break;
-    case "dotnet":
-      choices = [
-        { title: "6.0", value: "6.0" },
-        { title: "3.1", value: "3.1" },
-      ];
-      break;
-    case "dotnet-isolated":
-      choices = [{ title: "6.0", value: "6.0" }];
-      break;
-    default:
-      choices = [
-        { title: "16", value: "16" },
-        { title: "14", value: "14" },
-        { title: "12", value: "12" },
-      ];
-      break;
-  }
-  return choices;
 }
 
 function printFrameworkConfig(config: FrameworkConfig) {
